feat: add course icons to registration page
