+++ client/src/modules/Home/components/CourseSelector/index.tsx
@@ -1,8 +1,7 @@
 import { Select } from 'antd';
-import { PlayCircleTwoTone, CheckCircleTwoTone } from '@ant-design/icons';
 import React, { CSSProperties } from 'react';
 import { Course } from 'services/models';
-import { DEFAULT_COURSE_ICONS } from 'configs/course-icons';
+import { CourseIcon } from 'components/Icons/CourseIcon';
 
 type Props = {
   course: Course | null;
@@ -26,38 +25,10 @@ export function CourseSelector(props: Props) {
     >
       {courses.map(course => (
         <Select.Option style={getStatusCss(course)} key={course.id} value={course.id}>
-          {getStatusIcon(course)} {course.name} {getStatus(course)}
+          <CourseIcon course={course} /> {course.name} {getStatus(course)}
         </Select.Option>
       ))}
     </Select>
-  );
-}
-
-const getStatusIcon = (course: Course) => {
-  if (course.completed) {
-    return course.logo && DEFAULT_COURSE_ICONS[course.logo] ? (
-      <Logo url={DEFAULT_COURSE_ICONS[course.logo].archived} />
-    ) : (
-      <CheckCircleTwoTone twoToneColor="#aaa" />
-    );
-  }
-  return course.logo && DEFAULT_COURSE_ICONS[course.logo] ? (
-    <Logo url={DEFAULT_COURSE_ICONS[course.logo].active} />
-  ) : (
-    <PlayCircleTwoTone />
-  );
-};
-
-function Logo({ url }: { url: string }) {
-  return (
-    <>
-      <img width={16} height={16} src={url} className="img" />
-      <style jsx>{`
-        .img {
-          vertical-align: -0.125em;
-        }
-      `}</style>
-    </>
   );
 }
